refactor -> OpenQuant
